fix: add explicit ::int and ::numeric casts to PUT /api/order-items SQL
Without explicit type casts, PostgreSQL can fail when the JS number
parameter doesn't implicitly resolve to the correct numeric type for
the order_items columns. This caused 'Failed to update quantity' in
the purchase history page.

Also adds a regression test asserting the SQL uses \::int and
\::numeric.

+++ backend/src/routes/order-items.js
@@ -41,7 +41,7 @@ router.put('/api/order-items', orderItemsLimiter, requireAuth, asyncHandler(asyn
 
     const result = await client.query(
       `UPDATE order_items AS oi
-       SET quantity = $1, line_total = oi.unit_price * $1
+       SET quantity = $1::int, line_total = oi.unit_price * $1::numeric
        FROM orders AS o
        WHERE oi.id = $2
          AND oi.order_id = o.id
